+++ src/APIs/openweathermap/owmapi_types/Wind.java
@@ -3,6 +3,6 @@ package APIs.openweathermap.owmapi_types;
 public record Wind(
         float speed,
         int deg,
-        int gust
+        float gust
 ) {
 }
